refactor: move sample_wiki to wiki/mod.rs for consistent fixture centralization
Addresses review comment: import sample_wiki from list::tests_helper couples tests to list internals
