+++ src/cmd/wiki/delete.rs
@@ -46,7 +46,7 @@ pub fn delete_with(args: &WikiDeleteArgs, api: &dyn BacklogApi) -> Result<()> {
 mod tests {
     use super::*;
     use crate::api::wiki::Wiki;
-    use crate::cmd::wiki::list::tests_helper::sample_wiki;
+    use crate::cmd::wiki::sample_wiki;
     use anyhow::anyhow;
 
     struct MockApi {
