+++ src/cmd/wiki/show.rs
@@ -50,7 +50,7 @@ pub fn print_wiki(wiki: &Wiki) {
 mod tests {
     use super::*;
     use crate::api::wiki::Wiki;
-    use crate::cmd::wiki::list::tests_helper::sample_wiki;
+    use crate::cmd::wiki::sample_wiki;
     use anyhow::anyhow;
 
     struct MockApi {
