+++ src/cmd/wiki/update.rs
@@ -68,7 +68,7 @@ pub fn update_with(args: &WikiUpdateArgs, api: &dyn BacklogApi) -> Result<()> {
 mod tests {
     use super::*;
     use crate::api::wiki::Wiki;
-    use crate::cmd::wiki::list::tests_helper::sample_wiki;
+    use crate::cmd::wiki::sample_wiki;
     use anyhow::anyhow;
 
     struct MockApi {
